split project into multiple files
